pyproject.toml
[project]
name = "platform-engine-temp"
version = "0.1.0"
description = "Reusable workflow + Pulumi engine for provisioning services from platform.yaml"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.34.0",
    "pulumi>=3.0",
    "pulumi-aws>=6.0",
    "pulumi-cloudflare>=5.0",
    "pyyaml>=6.0",
]

[project.scripts]
platform = "devops.cli:main"
lint = "devops._scripts:lint"
lint-fix = "devops._scripts:lint_fix"
format = "devops._scripts:format"
type-check = "devops._scripts:type_check"
test = "devops._scripts:test"
test-cov = "devops._scripts:test_cov"

[project.optional-dependencies]
dev = [
    "pre-commit>=4.0.0",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "pyright>=1.1.406",
    "ruff>=0.8.4",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "N",
    "UP",
    "B",
    "SIM",
    "RUF",
    "I",
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["devops"]
force-sort-within-sections = true

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
reportMissingTypeStubs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["devops"]
omit = ["devops/__main__.py"]

[tool.coverage.report]
exclude_lines = ["pragma: no cover"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["devops"]
